Split out getEligibleGSuiteDomain
